Move everything under multipart file
Fix some nitpicks
